+++ src/PasswordManager.Application/Services/Abstractions/IAccountService.cs
@@ -4,19 +4,21 @@ using System.Threading.Tasks;
 
 namespace PasswordManager.Application.Services.Abstractions {
 	public interface IAccountService {
+
+		/// <param name="userId"></param>
 		/// <param name="accountName">Account Name to search. If null - count of all accounts will be returned.</param>
 		/// <returns>Count of all accouns found by <paramref name="accountName"/></returns>
-		Task<int> GetCountAsync(int userId, string accountName = null);
+		Task<int> GetCountAsync(long userId, string accountName = null);
 		/// <returns>Full Account info without password</returns>
-		Task<Account> GetFullAsync(int userId, long accountId);
+		Task<Account> GetFullAsync(long userId, long accountId);
 		/// <returns>List of basic Account info found by <paramref name="accountName"/></returns>
-		Task<IEnumerable<Account>> GetByNameAsync(int userId, int page, int pageSize, string accountName = null);
+		Task<IEnumerable<Account>> GetByNameAsync(long userId, int page, int pageSize, string accountName = null);
 		/// <returns>Password and its encryption state info</returns>
-		Task<Account> GetPasswordAsync(int userId, long accountId);
+		Task<Account> GetPasswordAsync(long userId, long accountId);
 		/// <returns>True on success</returns>
-		Task<bool> DeleteAccountAsync(int userId, long accountId);
-		Task<bool> AddAccountAsync(int userId, Account account);
-		Task UpdatePasswordAsync(int userId, long accountId, string password, bool encrypted);
+		Task<bool> DeleteAccountAsync(long userId, long accountId);
+		Task<bool> AddAccountAsync(long userId, Account account);
+		Task UpdatePasswordAsync(long userId, long accountId, string password, bool encrypted);
 		Task UpdateAccountAsync();
 	}
 }
