+++ src/PasswordManager.Bot/Services/Abstractions/IPasswordDecryptionService.cs
@@ -5,13 +5,13 @@ namespace PasswordManager.Bot.Services.Abstractions {
 		/// <summary>
 		/// Saves account with encrypted password of user to return it for decryption later
 		/// </summary>
-		void StartDecryptionRequest(int userId, Account account);
+		void StartDecryptionRequest(long userId, Account account);
 		/// <summary>
 		/// Releases held account
 		/// </summary>
-		void FinishDecryptionRequest(int userId);
+		void FinishDecryptionRequest(long userId);
 		/// <summary></summary>
 		/// <returns>Account with encrypted password of user or <see langword="null"/> if there is no decryption request</returns>
-		Account GetAccount(int userId);
+		Account GetAccount(long userId);
 	}
 }
