+++ src/PasswordManager.Bot/Services/Abstractions/IPasswordEncryptionService.cs
@@ -4,13 +4,13 @@ namespace PasswordManager.Bot.Services.Abstractions {
 		/// <summary>
 		/// Saves accountId of user to return it for recryption later
 		/// </summary>
-		void StartEncryptionRequest(int userId, long accountId);
+		void StartEncryptionRequest(long userId, long accountId);
 		/// <summary>
 		/// Releases held accountId
 		/// </summary>
-		void FinishEncryptionRequest(int userId);
+		void FinishEncryptionRequest(long userId);
 		/// <summary></summary>
 		/// <returns>accountId of user or <see langword="null"/> if there is no encryption request</returns>
-		long? GetAccountId(int userId);
+		long? GetAccountId(long userId);
 	}
 }
